requirements.txt
pandas
bcrypt
CacheControl==0.12.5
certifi==2018.4.16
chardet==3.0.4
click==6.7
Cython==0.28.4
Flask==1.0.2
Flask-Cors==3.0.6
flask-swagger-ui==3.6.0
future==0.16.0
h5py==2.8.0
idna==2.7
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
lockfile==0.12.2
MarkupSafe==1.0
mistune==0.8.3
msgpack==0.5.6
numpy==1.15.0
passlib==1.7.1
pip==18.0
PyJWT==1.6.4
pyparsing==2.2.0
PyYAML==3.13
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.19.1
setuptools==40.0.0
shellescape==3.4.1
six==1.11.0
typing==3.6.4
urllib3==1.23
uWSGI==2.0.17.1
Werkzeug==0.14.1
wheel==0.31.1
xkcdpass==1.16.5
